requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
discord.py==2.6.3
Flask==3.1.2
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
yt-dlp==2025.9.26
